+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Flag for deny split transaction
+
 ## 2.12.3 ##
 * Add six package to requirements
 * Fixed error while passing date parameter in execute